requirements.txt
Flask
numpy
tflite-runtime
pillow
